+++ src/cores/metas/datagroup_meta.py
@@ -5,10 +5,10 @@ from typing import Optional
 from interfaces.meta_interface import MetaInterface
 
 @dataclass
-class DataUnitMeta(MetaInterface):
+class DataGroupMeta(MetaInterface):
     name: str
     path: Path
-    created_at: datetime = field(default_factory=lambda:datetime.now()) # 初期値を現在の時間にする
+    created_at: Optional[datetime] = field(default_factory=lambda:datetime.now()) # 初期値を現在の時間にする
 
     def __post_init__(self):
         if not self.name: