+++ tests/cores/test_dataunit.py
@@ -10,8 +10,7 @@ def test_create_dataunit_success():
     metadata = DataUnitMeta(
         name=name,
         path=Path('./aaa.csv'),
-        created_at=datetime.now(),
-        format='csv'
+        created_at=datetime.now()
     )
     unit = DataUnit(df, metadata)
     assert unit.name == 'unit1'
@@ -24,8 +23,7 @@ def test_create_dataunit_empty_df_raises():
     metadata = DataUnitMeta(
         name=name,
         path=Path('./aaa.csv'),
-        created_at=datetime.now(),
-        format='csv'
+        created_at=datetime.now()
     )
     with pytest.raises(ValueError):
         DataUnit(df, metadata)
@@ -36,8 +34,7 @@ def test_with_updated_name():
     metadata = DataUnitMeta(
         name=name,
         path=Path('./aaa.csv'),
-        created_at=datetime.now(),
-        format='csv'
+        created_at=datetime.now()
     )
     unit = DataUnit(df, metadata)
     new_name = 'new_unit'
@@ -45,38 +42,3 @@ def test_with_updated_name():
     assert new_unit.df.equals(df)
     assert new_unit.metadata.name == new_name
     assert unit.metadata.name == "unit1"  # イミュータブル確認
-
-def test_with_updated_df():
-    df1 = pd.DataFrame({'a': [1]})
-    df2 = pd.DataFrame({'b': [2]})
-    name = 'unit1'
-    metadata = DataUnitMeta(
-        name=name,
-        path=Path('./aaa.csv'),
-        created_at=datetime.now(),
-        format='csv'
-    )
-    unit = DataUnit(df1, metadata)
-    new_unit = unit.with_updated_df(df2)
-    assert new_unit.df.equals(df2)
-    assert unit.df.equals(df1) # イミュータブル確認
-
-def test_with_updated_metadata():
-    df = pd.DataFrame({'a': [1]})
-    
-    name_1 = 'unit1'
-    metadata_1 = DataUnitMeta(
-        name=name_1,
-        path=Path('./aaa.csv'),
-        created_at=datetime.now(),
-        format='csv'
-    )
-    unit = DataUnit(df, metadata_1)
-
-    new_unit = unit.with_update_metadata("format", "json")
-    assert new_unit.metadata.format == "json"
-
-    assert unit.metadata == metadata_1  # イミュータブル確認
-
-    with pytest.raises(AttributeError):
-        new_unit = unit.with_update_metadata("unknown_key", 'hello')